Updated code to rename variables.
One location identifier in this code now. Used by the pump and voltage functions. Next need to extract further and place in config file and open that here and also use with flower power bridge. Need to next test this code. Also need to try running flower bridge script from this? Will this code be running constantly? Or just be called every so often by a cron job?