Created master.py
Needs testing. 
Made 1 logging function that logs to internet or if this fails to a file or if this fails prints out data
Started trying to catch import errors and log this
Need to add in code to call shell script for flower power. Need to know where the flower power folder will be? Assume inside this git repo? 
Need to add in code to schedule function call to pumpWatering function at the next watering time? Look into "scheduling a function at a certain time python"? 
Need to check if master.py is already running on the computer or not..? Only want 1 instance of the program (since cron will rerun master.py every so often?)